\name{AcceProximal-package}
\alias{AcceProximal-package}
\alias{AcceProximal}
\docType{package}
\title{
\packageTitle{AcceProximal}
}
\description{
\packageDescription{AcceProximal}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{AcceProximal}
\packageIndices{AcceProximal}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{AcceProximal}

Maintainer: \packageMaintainer{AcceProximal}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
